Apply @reduxjs/toolkit
Install @reduxjs/toolkit
Uninstall redux/thunk
Replace actions, reducer with slice
