+++ package.json
@@ -42,11 +42,11 @@
   },
   "dependencies": {
     "@codeceptjs/configure": "^0.5.2",
+    "@reduxjs/toolkit": "^1.4.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.2.0",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0"
+    "redux": "^4.0.5"
   }
 }
